refactored init and signature of TweezerPlotter and renamed as TweezerStatistician to reflect the fact that it is doing several >preprocessing analysis tasks in addition to plotting. Started adding framework in TweezerStatistician to begin using MLOOP again, by saving analysis results to the shot_h5 file. Deleted analysis files using old code.
